refactor: componentização de funções e constantes, e, ajuste na identação de arquivos #14
