+++ src/components/DadosPessoaisForm.tsx
@@ -1,0 +1,98 @@
+import React from 'react';
+import { InputField } from './InputField';
+
+interface DadosPessoaisFormProps {
+    formData: {
+        nome: string;
+        dob: string;
+        cpf: string;
+        telefone: string;
+        email: string;
+        senha: string;
+        confirmacaoSenha: string;
+    };
+    errors: { [key: string]: string };
+    onChange: (field: string, value: string) => void;
+}
+
+export const DadosPessoaisForm: React.FC<DadosPessoaisFormProps> = ({ formData, errors, onChange }) => {
+    return (
+        <>
+            <h2 style={{ marginTop: '2rem', marginBottom: '1rem', paddingTop: '0.5rem' }}>
+                Dados Pessoais</h2>
+            <InputField
+                id="nome"
+                label="Nome Completo"
+                type="text"
+                value={formData.nome}
+                placeholder="Digite seu nome"
+                onChange={(e) => onChange('nome', e.target.value)}
+                required
+                errorMessage={errors.nome}
+            />
+            <div className="grid grid-cols-2 gap-4">
+                <InputField
+                    id="dob"
+                    label="Data de Nascimento"
+                    type="date"
+                    value={formData.dob}
+                    onChange={(e) => onChange('dob', e.target.value)}
+                    required
+                    errorMessage={errors.dob}
+                />
+                <InputField
+                    id="cpf"
+                    label="CPF"
+                    type="text"
+                    value={formData.cpf}
+                    placeholder="Digite seu CPF"
+                    onChange={(e) => onChange('cpf', e.target.value)}
+                    required
+                    errorMessage={errors.cpf}
+                />
+            </div>
+            <InputField
+                id="telefone"
+                label="Telefone"
+                type="text"
+                value={formData.telefone}
+                placeholder="Digite seu telefone"
+                onChange={(e) => onChange('telefone', e.target.value)}
+                required
+                errorMessage={errors.telefone}
+            />
+            <InputField
+                id="email"
+                label="E-mail"
+                type="email"
+                value={formData.email}
+                placeholder="Digite seu e-mail"
+                onChange={(e) => onChange('email', e.target.value)}
+                required
+                errorMessage={errors.email}
+            />
+            <div className="grid grid-cols-2 gap-4">
+                <InputField
+                    id="senha"
+                    label="Senha"
+                    type="password"
+                    value={formData.senha}
+                    placeholder="Digite sua senha"
+                    onChange={(e) => onChange('senha', e.target.value)}
+                    required
+                    errorMessage={errors.senha}
+                />
+                <InputField
+                    id="confirmacaoSenha"
+                    label="Confirmação de Senha"
+                    type="password"
+                    value={formData.confirmacaoSenha}
+                    placeholder="Confirme sua senha"
+                    onChange={(e) => onChange('confirmacaoSenha', e.target.value)}
+                    required
+                    errorMessage={errors.confirmacaoSenha}
+                />
+            </div>
+        </>
+    );
+};
