+++ src/components/InputField.tsx
@@ -33,9 +33,8 @@ export const InputField: React.FC<InputFieldProps> = ({
       <input
         type={type}
         id={id}
-        className={`mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 ${
-          errorMessage ? 'border-red-500' : 'border-gray-700'
-        } hover:border-black`}
+        className={`mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 ${errorMessage ? 'border-red-500' : 'border-gray-700'
+          } hover:border-black`}
         placeholder={placeholder}
         value={value}
         onChange={onChange}
@@ -44,8 +43,8 @@ export const InputField: React.FC<InputFieldProps> = ({
         style={{
           ...style,
           height: '2.5rem',
-          border: `1px solid ${errorMessage ? 'red' : 'currentColor'}`, // Borda na cor do texto
-          color: 'currentColor', // Garante que o texto e a borda tenham a mesma cor
+          border: `1px solid ${errorMessage ? 'red' : 'currentColor'}`,
+          color: 'currentColor',
           padding: '0.5rem',
         }}
       />
