+++ src/index.css
@@ -10,25 +10,29 @@ h2 {
     @apply leading-relaxed;
 }
 
-select{
+select {
     margin-bottom: 1rem;
-    padding: 0 !important;    
+    padding: 0 !important;
     border: 1px solid #e2e8f0;
     border-radius: 0.375rem;
     width: 100%;
     height: 2rem;
     outline: none;
-    &:focus{
+
+    &:focus {
         border-color: #3182ce;
     }
-    option{
+
+    option {
         background-color: #f9fafc;
         color: #63666a;
     }
-    option:hover{
+
+    option:hover {
         background-color: #e9ecef;
     }
-    option:selected{
+
+    option:selected {
         background-color: #3182ce;
         color: #fff;
     }
