+++ src/types/denuncia.ts
