+++ src/util/validarTelefone.ts
@@ -1,4 +1,4 @@
 export function validarTelefone(celular: string): boolean {
-    const celularLimpo = celular.replace(/\D/g, ''); // Remove caracteres não numéricos
-    return celularLimpo.length >= 10; // Verifica se tem pelo menos 10 dígitos
-  }
+  const celularLimpo = celular.replace(/\D/g, '');
+  return celularLimpo.length >= 10;
+}
